+++ History.md
@@ -5,11 +5,17 @@ History & Notes
 ================
 Note: Simple documentation updates are not listed here.
 
+#### 0.4.0 / 2015-10-09
+
+* Changed: `options.createGetterOn` and `options.createSetterOn` are renamed as `options.getter` and `options.setter`.
+* Added: `options.createAccessorsIfNotExists` added.
+* Fixed: Wrong path parameter caused reply called twice. Fixed. 
+
 #### 0.3.0 / 2015-10-07
 * Added: Cookie support. `options.order` to change order of process to determine locale. It is possible to proritize query etc. over url parameters now. 
 
 #### 0.2.1 / 2015-10-07
-* Changed: `options.createGetter` and `options.createSetter` are renamed as options.createGetterOn and options.createSetterOn
+* Changed: `options.createGetter` and `options.createSetter` are renamed as `options.createGetterOn` and `options.createSetterOn`
 
 #### 0.2.0 / 2015-10-07
 * Added: `getDefaultLocale()` exposed function.